+++ resym/src/main.rs
@@ -11,7 +11,6 @@ use memory_logger::blocking::MemoryLogger;
 use resym_app::ResymApp;
 
 const PKG_NAME: &str = env!("CARGO_PKG_NAME");
-const PKG_VERSION: &str = env!("CARGO_PKG_VERSION");
 
 fn main() -> Result<()> {
     let logger = MemoryLogger::setup(log::Level::Info)?;